Split some functionality to their own files
